move patches from this project to hts-engine to separate source file
